Journal-part2/requirements.txt
# Gemini Engine Components 6 & 7 - Production Dependencies
# Core AI and ML libraries
google-generativeai>=0.3.0
openai>=1.0.0
tiktoken>=0.5.0

# Data processing and validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async and HTTP
aiohttp>=3.8.0
httpx>=0.24.0
asyncio-mqtt>=0.11.0

# Database and caching
redis>=4.5.0
asyncpg>=0.28.0
pinecone-client>=2.2.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0

# Monitoring and logging
structlog>=23.0.0
prometheus-client>=0.17.0

# Testing and development
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Performance and optimization
orjson>=3.8.0
ujson>=5.7.0

# Security
cryptography>=41.0.0
passlib>=1.7.0

# Date and time
python-dateutil>=2.8.0
pytz>=2023.3

# Text processing
nltk>=3.8.0
spacy>=3.6.0

# Vector operations
numpy>=1.24.0
scipy>=1.10.0

# Machine learning utilities
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0

# API and web framework
fastapi>=0.100.0
uvicorn>=0.22.0

# Background tasks
celery>=5.3.0
redis>=4.5.0

# Metrics and monitoring
prometheus-client>=0.17.0
statsd>=4.0.0

# Error tracking
sentry-sdk>=1.28.0

# Development tools
ipython>=8.0.0
jupyter>=1.0.0


# FastAPI Integration Requirements
# Gemini Engine API Dependencies

# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS and middleware
python-cors==1.7.0

# Environment and configuration
python-dotenv==1.0.0

# Async support
asyncio-mqtt==0.16.1

# HTTP client
httpx==0.25.2
requests==2.31.0

# JSON handling
orjson==3.9.10

# Logging and monitoring
structlog==23.2.0

# Date and time handling
python-dateutil==2.8.2

# Type hints and validation
typing-extensions==4.8.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: API documentation enhancements
fastapi-swagger-ui==0.1.0
